+++ Examples/ApiKey/Program.cs
@@ -2,7 +2,11 @@ using ChustaSoft.Auth.ApiKey;
 
 var builder = WebApplication.CreateBuilder(args);
 
-builder.Services.ConfigureApiKeyAuthentication((token) => { return token == "Test_ApiKey"; });
+// Option 1: Using the extension method with a simple lambda function as the token provider
+//builder.Services.ConfigureApiKeyAuthentication((token) => { return token == "Test_ApiKey"; });
+
+// Option 2: Using the extension method with a custom token provider class that implements the ITokenProvider interface
+builder.Services.ConfigureApiKeyAuthentication<InternalTokenProvider>();
 
 var app = builder.Build();
 
@@ -30,3 +34,12 @@ app.MapGet("/weatherforecast", () =>
 }).RequireAuthorization();
 
 app.Run();
+
+
+public class InternalTokenProvider : ITokenProvider
+{
+    public bool Validate(string token)
+    {
+        return token == "Test_ApiKey";
+    }
+}