+++ Examples/Basic/Program.cs
@@ -1,9 +1,16 @@
+using ChustaSoft.Auth.ApiKey;
 using ChustaSoft.Auth.Basic;
 
 var builder = WebApplication.CreateBuilder(args);
 
 builder.Services.AddControllers();
-builder.Services.ConfigureBasicAuthentication((username, password) => { return username == "test" && password == "pass"; });
+
+// Option 1: Using the extension method with a simple lambda function as the credentials provider
+//builder.Services.ConfigureBasicAuthentication((username, password) => { return username == "test" && password == "pass"; });
+
+// Option 2: Using the extension method with a custom credentials provider class that implements the ICredentialsProvider interface
+builder.Services.ConfigureBasicAuthentication<InternalCredentialsProvider>();
+
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 
@@ -24,3 +31,12 @@ app.UseAuthorization();
 app.MapControllers();
 
 app.Run();
+
+
+public class InternalCredentialsProvider : ICredentialsProvider
+{
+    public bool Validate(string username, string password)
+    {
+        return username == "test" && password == "pass";
+    }
+}