+++ NuGet/ChustaSoft.Auth.Basic/ICredentialsProvider.cs
@@ -1,0 +1,6 @@
+﻿namespace ChustaSoft.Auth.ApiKey;
+
+public interface ICredentialsProvider
+{
+    bool Validate(string username, string password);
+}