+++ NuGet/ChustaSoft.Auth.Basic/changelog.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Releases]
 
-## [1.0.0] - XXXX-XX-XX
+## [1.0.0] - 2026-06-02
 ### Added
-- [TBD]
+- Added AuthenticationHandler mechanism for Basic Auth via lambda function as provider
+- Added generic ConfigureBasicAuthentication method with ICredentialsProvider support
+- .NET10 support